requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
openai==1.6.1
icalendar==5.0.11
requests==2.31.0
python-dateutil==2.8.2
Werkzeug==3.0.1
